Fixed button animation bug
I improved the button's animation when you hover and not hover the button. Before when you would hover and not hover fast or mix them up when the animation was not fully complete, the animation would look buggy since the not hover animation always started at scale(1.1), rotate(10deg), and font size of 25px instead of using the button's current scale, rotation, and font size.

What I did was saved the button's current scale, rotation, and font size, so whenever you hover the button, the button instead of going from a fixed value to the target value, it goes from the current values to the target value.

When you not hover, instead of going from a the target fixed value (1.2 scale, 5 rotation degree, and 25 font size) to the default values (scale 1, rotation 0 degrees, font size 18px), the animation goes from the current values to the default values.

Previously the target scale was 1.1, and the target rotation degree was 10, but I changed the target scale to 1.2, and the rotation degree to 5. The font size target being 25px is the same, that did not change.