Fixed spam bug
Before you could keep asking questions which would cause the 8 ball to do multiple shakes which would call the shake function and the animation timeouts and intervals which was a bug since the previous 8 ball response would display and then quickly the next one would, and this would repeat for how many times you asked the question before one finished.

The change I made is so that you have to wait for the magic 8 ball to display text first before you can submit your next question.

Whenever you submit a question the input field becomes empty. However, if you try to submit a question while the magic 8 ball is not finished giving it's answer, the input field will not become empty. This is to let the user know their question has not been submitted yet.